Sart refactor duplicate classes
